Refactorred playlist class.
We add a new tracklist class, and use it for storing
tracks in playlist. This class will be useful next,
for storing more information about the added class.
For example, if we want to store the file path, we
can just add a new property in this class, and we
won't clutter the playlist class.