Create track class.
We now store all files in an array (tracks), with
every track stored as an object. We can use this
object later to attach it into an audio element and
play the song.

The "id" attribute on track object is useful to know
which music we'll be playing, and which we will play
next. Hopefully, we will also found another use of it
in the future (thinking of maybe shuffle).